Fix another win32 test failure.
As of this commit, the entire test suite passes on a system that only has:

- Python 2.6.4 from Python.org
- Pyreadline 1.5
- IPython (this branch)
- nose 0.11.1

Tests that require Twisted, graphics or other machinery get skipped, but
everything else runs and passes.

The glaring remaining problem on win32 is that right now, colored output is
broken for some reason (though prompts are OK).  I'll try to fix that next.